updater-drivers: add new package
The idea behind this package is that we want to install automatically
some optional drivers when user connects new device. That of course does
not mean immediatelly but rather with next update but installing
drivers automatically (and only those required) is much needed feature
for end users.
The implementation only constructs list of devices connected to router.
This list is exported as table and dedicated list is called. That list
uses that table to request appropriate drivers.
Using list over managing it locally on router it cleaner for us as we
can reuse same list for bootstrap and other tasks (not need to replan to
generate medkit).

Currently we support USB devices and PCI devices. In future we can
potentially add any other bus with discovery functionality to include
drivers for devices on those as well.