tos3to4: fix interface assigment for no interface
For some protocols there is fully valid setting with no ifname but state
configuration injected by netifd inserts ifname. Original approach was
to take devices (the variable used for original ifname value in state)
and if those were not present then ifname. The idea was that if there
were no devices then device was down or not configured and ifname
contains real configuration. That is unfortunately not true clearly in
case of protocols such as wireguard.
New approach is cleaner as we just simply ignore state and read real
configuration directly. In the end this even simplifies the code.