lighttpd: add patch for meson.lua
With this patch, lighttpd meson build will look for liblua.so and lua/lua.h.
With this patch, lighttpd meson build does not look for lua w/ version, e.g. liblua-5.3.so and lua-5.3/lua.h

lighttpd does not have requirements on a specific lua version and should work with any of them, so the openwrt packaging should end up detecting that the lighttpd build is satisfied by the declared dependeny on liblua package in the lighttpd/Makefile.

This is going to fix some lighttpd modules which were not compiled.

Signed-off-by: Glenn Strauss <gstrauss@gluelogic.com>