lighttpd: remove mod_auth_ticket
It was added in this commit: 88ef23faf2f603ba590f7f902a4558d7df494204
but Michal was experimenting with ticket mechanism, but never got to it.
At least this package is not installed by default, but meant to be just
for developers.

Auth_ticket is not incompatible with Lighttpd version 1.4.56
Lighttpd developer suggested to cosider and rather use lighttpd-mod-authn-tkt.

Fixes: 6d1a4393f22f9a5f775afc4da219f47bfdb40226